+++ packages/components-native/src/Form/components/FormMessage/components/InternalFormMessage/messages.ts
@@ -1,0 +1,10 @@
+import { defineMessages } from "react-intl";
+
+export const messages = defineMessages({
+  goBackButton: {
+    id: "goBackButton",
+    defaultMessage: "Go Back",
+    description:
+      "The label for the fallback secondary button when it doesn't exist",
+  },
+});