+++ packages/components-native/src/Menu/components/Overlay/Overlay.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+import { Pressable, View } from "react-native";
+import { styles } from "./Overlay.style";
+import { OverlayProp } from "../../types";
+
+export function Overlay({ setOpen }: OverlayProp): JSX.Element {
+  return (
+    <Pressable
+      onPressIn={() => {
+        setOpen(false);
+      }}
+    >
+      <View style={styles.overlay} />
+    </Pressable>
+  );
+}