+++ packages/components/src/Menu/Menu.tsx
@@ -12,6 +12,7 @@ import { v1 as uuidv1 } from "uuid";
 import classnames from "classnames";
 import { AnimatePresence, motion } from "framer-motion";
 import { useOnKeyDown } from "@jobber/hooks/useOnKeyDown";
+import { useRefocusOnActivator } from "@jobber/hooks/useRefocusOnActivator";
 import { IconNames } from "@jobber/design";
 import styles from "./Menu.css";
 import { Button } from "../Button";
@@ -92,6 +93,7 @@ export function Menu({ activator, items }: MenuProps) {
       setPosition(newPosition);
     }
   }, [visible, fullWidth]);
+  useRefocusOnActivator(visible);
 
   if (!activator) {
     activator = (
@@ -117,7 +119,11 @@ export function Menu({ activator, items }: MenuProps) {
   });
 
   return (
-    <div className={wrapperClasses} ref={wrapperRef}>
+    <div
+      className={wrapperClasses}
+      ref={wrapperRef}
+      onClick={handleParentClick}
+    >
       {React.cloneElement(activator, {
         onClick: toggle(activator.props.onClick),
         id: buttonID,
@@ -196,6 +202,14 @@ export function Menu({ activator, items }: MenuProps) {
     event.stopPropagation();
     key === "Escape" && hide();
   }
+
+  function handleParentClick(event: MouseEvent<HTMLDivElement>) {
+    // Since the menu is being rendered within the same parent as the activator,
+    // we need to stop the click event from bubbling up. If the Menu component
+    // gets added within a parent that has a click handler, any click on the
+    // menu will trigger the parent's click handler.
+    event.stopPropagation();
+  }
 }
 
 interface SectionHeaderProps {
@@ -255,8 +269,9 @@ function Action({
   const actionButtonRef = useRef() as RefObject<HTMLButtonElement>;
 
   useEffect(() => {
-    if (actionButtonRef.current && shouldFocus) {
-      actionButtonRef.current.focus();
+    if (shouldFocus) {
+      // Focus on the next tick to allow useRefocusOnActivator to initialize
+      setTimeout(() => actionButtonRef.current?.focus(), 0);
     }
   }, [shouldFocus]);